Move interfaces and interface tests out of the batch folder
