+++ pennylane/__init__.py
@@ -86,7 +86,7 @@ from pennylane.debugging import snapshots
 from .collections import QNodeCollection, dot, map, sum
 import pennylane.grouping  # pylint:disable=wrong-import-order
 import pennylane.gradients  # pylint:disable=wrong-import-order
-from pennylane.interfaces.batch import execute  # pylint:disable=wrong-import-order
+from pennylane.interfaces import execute  # pylint:disable=wrong-import-order
 
 # Look for an existing configuration file
 default_config = Configuration("config.toml")
