+++ tests/interfaces/test_autograd.py
@@ -22,7 +22,7 @@ from pennylane import numpy as np
 
 import pennylane as qml
 from pennylane.gradients import finite_diff, param_shift
-from pennylane.interfaces.batch import execute
+from pennylane.interfaces import execute
 
 
 class TestAutogradExecuteUnitTests:
@@ -35,7 +35,7 @@ class TestAutogradExecuteUnitTests:
         mock.side_effect = ImportError()
 
         try:
-            del sys.modules["pennylane.interfaces.batch.autograd"]
+            del sys.modules["pennylane.interfaces.autograd"]
         except:
             pass
 
@@ -206,7 +206,7 @@ class TestCaching:
     def test_cache_maxsize(self, mocker):
         """Test the cachesize property of the cache"""
         dev = qml.device("default.qubit", wires=1)
-        spy = mocker.spy(qml.interfaces.batch, "cache_execute")
+        spy = mocker.spy(qml.interfaces, "cache_execute")
 
         def cost(a, cachesize):
             with qml.tape.QuantumTape() as tape:
@@ -227,7 +227,7 @@ class TestCaching:
     def test_custom_cache(self, mocker):
         """Test the use of a custom cache object"""
         dev = qml.device("default.qubit", wires=1)
-        spy = mocker.spy(qml.interfaces.batch, "cache_execute")
+        spy = mocker.spy(qml.interfaces, "cache_execute")
 
         def cost(a, cache):
             with qml.tape.QuantumTape() as tape:
