+++ tests/interfaces/test_torch.py
@@ -21,7 +21,7 @@ torch = pytest.importorskip("torch")
 
 import pennylane as qml
 from pennylane.gradients import finite_diff, param_shift
-from pennylane.interfaces.batch import execute
+from pennylane.interfaces import execute
 
 
 class TestTorchExecuteUnitTests:
@@ -155,7 +155,7 @@ class TestCaching:
     def test_cache_maxsize(self, mocker):
         """Test the cachesize property of the cache"""
         dev = qml.device("default.qubit", wires=1)
-        spy = mocker.spy(qml.interfaces.batch, "cache_execute")
+        spy = mocker.spy(qml.interfaces, "cache_execute")
 
         def cost(a, cachesize):
             with qml.tape.QuantumTape() as tape:
@@ -179,7 +179,7 @@ class TestCaching:
     def test_custom_cache(self, mocker):
         """Test the use of a custom cache object"""
         dev = qml.device("default.qubit", wires=1)
-        spy = mocker.spy(qml.interfaces.batch, "cache_execute")
+        spy = mocker.spy(qml.interfaces, "cache_execute")
 
         def cost(a, cache):
             with qml.tape.QuantumTape() as tape:
